requirements.txt
Flask==3.1.0
requests==2.32.3
flask_sqlalchemy==3.1.1
flask_migrate==4.0.7
flask_cors==5.0.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6  # No newer version found in your list, keeping original
python-dotenv==1.0.1
runpod==1.7.7
psycopg2-binary